Update 3.2.0 CLI Overhaul
