+++ src/config/__init__.py
@@ -1,0 +1,3 @@
+from .config import *
+from .parsersettings import *
+from .paths import *
