+++ src/config/netwatch.ini
@@ -1,15 +1,21 @@
 # Start Main Menu Configurations
 [general_config]
-__version__ = 3.1.4
+__version__ = 3.2.0
 prompt = netwatch ~#
 config_file_path = ./config.ini
 logger_config_path = ./logger.yaml
 tooldir = ../tools/
-logdir = ../logs/
+scandir = ../scans/
 storehistory = false
 ovpn_path = /path/to/ovpn/file.ovpn
 menu_data_path = /menuData.json
 module_logos_path = /moduleLogos.py
+
+[DefaultTheme]
+primary = white
+warning = yellow
+error = red
+critical = red [bold]
 
 [netwatch]
