+++ src/logger/__init__.py
@@ -1,0 +1,2 @@
+from .color_formatter import *
+from .logger import *
