+++ src/modules/__init__.py
@@ -1,0 +1,8 @@
+from .colorhelper import *
+from .constants import *
+from .notify import *
+from .termutils import *
+from .packagemanager import *
+from .updatehandler import *
+from .tablecreator import *
+from .sites import *
