+++ src/modules/termutils.py
@@ -167,186 +167,3 @@ class TextOutputEntry():
             return "%s%s%s" % (self.color, self.text, colorama.Style.RESET_ALL)
         else:
             return self.text
-
-
-__all__ = ['c', 'use_prop']
-
-# Initialization:
-colorama.just_fix_windows_console()
-
-
-def c(text: str) -> str:
-    result = text
-
-    # Colors:
-    theme = get_theme()
-
-    for key, color in theme.colors.items():
-        styled_portion = re.findall(f"<{key}>(.*?)</{key}>", text)
-
-        if not styled_portion:
-            continue
-
-        for portion in styled_portion:
-            result = re.sub(f"<{key}>(.*?)</{key}>",
-                            colored(portion, color=color),
-                            result, 1)
-
-    # Highlights
-    for key, highlight in theme.highlights.items():
-        styled_portion = re.findall(f"<{key}>(.*?)</{key}>", text)
-
-        if not styled_portion:
-            continue
-
-        for portion in styled_portion:
-            result = re.sub(f"<{key}>(.*?)</{key}>",
-                            colored(portion, on_color=highlight),
-                            result, 1)
-
-    # Attributes:
-    for key, attribute in theme.attributes.items():
-        styled_portion = re.findall(f"<{key}>(.*?)</{key}>", text)
-
-        if not styled_portion:
-            continue
-
-        for portion in styled_portion:
-            result = re.sub(f"<{key}>(.*?)</{key}>",
-                            colored(portion, attrs=attribute),
-                            result, 1)
-
-    # Custom Themes:
-    for key, style in theme.content.items():
-        styled_portion = re.findall(f"<{key}>(.*?)</{key}>", text)
-
-        if not styled_portion:
-            continue
-
-        for portion in styled_portion:
-            result = re.sub(f"<{key}>(.*?)</{key}>",
-                            colored(portion, color=style.color,
-                                    on_color=style.highlight, attrs=style.attributes),
-                            result, 1)
-
-    return result
-
-
-def use_prop(text: str, prop: str) -> str:
-    style = get_theme().content.get(prop)
-
-    text = f"<{style.color}>{text}</{style.color}>"
-
-    if style.highlight:
-        text = f"<{style.highlight}>{text}</{style.highlight}>"
-
-    if style.attributes:
-        for attribute in style.attributes:
-            text = f"<{attribute}>{text}</{attribute}>"
-
-    return c(text)
-
-
-def primary(text: str) -> str:
-    return c(f"<primary>{text}</primary>")
-
-
-def warning(text: str) -> str:
-    return c(f"<warning>{text}</warning>")
-
-
-def error(text: str) -> str:
-    return c(f"<error>{text}</error>")
-
-
-def critical(text: str) -> str:
-    return c(f"<critical>{text}</critical>")
-
-# Style:
-
-
-@dataclass
-class Style:
-    color: str = ''
-    highlight: str = ''
-    attributes: list[str] = field(default_factory=list)
-
-
-@dataclass
-class Theme:
-    content: dict = field(default_factory=dict)
-
-    colors: dict = field(default_factory=dict)
-    highlights: dict = field(default_factory=dict)
-    attributes: dict = field(default_factory=dict)
-
-    def __post_init__(self):
-        self.colors = {
-            'grey': 'grey',
-            'red': 'red',
-            'green': 'green',
-            'yellow': 'yellow',
-            'blue': 'blue',
-            'magenta': 'magenta',
-            'cyan': 'cyan',
-            'white': 'white'
-        }
-
-        self.highlights = {
-            'on_grey': 'on_grey',
-            'on_red': 'on_red',
-            'on_green': 'on_green',
-            'on_yellow': 'on_yellow',
-            'on_blue': 'on_blue',
-            'on_magenta': 'on_magenta',
-            'on_cyan': 'on_cyan',
-            'on_white': 'on_white'
-        }
-
-        self.attributes = {
-            'bold': 'bold',
-            'dark': 'dark',
-            'underline': 'underline',
-            'blink': 'blink',
-            'reverse': 'reverse',
-            'concealed': 'concealed'
-        }
-
-
-theme = Theme()
-
-
-def get_theme() -> Theme():
-    return theme
-
-
-def set_theme(theme_name: str):
-    global theme
-    theme = load_theme(theme_name)
-
-
-def load_theme(theme_name: str) -> Theme:
-    theme_section = read_section(theme_name)
-    theme_content = dict(theme_section)
-
-    content = {}
-    for key, item in theme_content.items():
-        style = Style()
-
-        color = re.sub("[\{\[].*?[\}\]]", "", item).strip()
-
-        highlight = re.findall('{(.*?)}', item)
-        if highlight:
-            highlight = highlight[0]
-
-        attributes = ''.join(re.findall('\[(.*?)\]', item)).split(' ')
-        if all(x.isspace() or not x for x in attributes):
-            attributes = None
-
-        style.color = color if color else 'white'
-        style.highlight = highlight if highlight else None
-        style.attributes = attributes if attributes else None
-
-        content[key] = style
-
-    return Theme(content)
